# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TGMSIDC Equipment Procurement and Order Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rate-contracts:
    get:
      operationId: listRateContracts
      tags: [rate-contracts]
      summary: List all rate contracts
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: equipmentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of rate contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RateContract"
    post:
      operationId: createRateContract
      tags: [rate-contracts]
      summary: Create a new rate contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRateContractBody"
      responses:
        "201":
          description: Created rate contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateContract"
  /rate-contracts/{id}:
    get:
      operationId: getRateContract
      tags: [rate-contracts]
      summary: Get a rate contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rate contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateContract"
        "404":
          description: Not found
    patch:
      operationId: updateRateContract
      tags: [rate-contracts]
      summary: Update a rate contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRateContractBody"
      responses:
        "200":
          description: Updated rate contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateContract"
...
  /tenders:
    get:
      operationId: listTenders
      tags: [tenders]
      summary: List all tenders
      responses:
        "200":
          description: List of tenders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tender"
    post:
      operationId: createTender
      tags: [tenders]
      summary: Create a tender record for an indent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenderBody"
      responses:
        "201":
          description: Created tender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"
  /tenders/{id}:
    get:
      operationId: getTender
      tags: [tenders]
      summary: Get a tender
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tender details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"
    patch:
      operationId: updateTender
      tags: [tenders]
      summary: Update tender milestone status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenderBody"
      responses:
        "200":
          description: Updated tender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"

  # ─── DELIVERIES ─────────────────────────────────────────────────────────────
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Register a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorBody"
      responses:
        "201":
          description: Created vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found

  # ─── INSTITUTIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RateContract:
      type: object
      properties:
        id:
          type: integer
        contractNumber:
          type: string
        equipmentId:
          type: integer
        equipmentName:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        unitPrice:
          type: number
        gstRate:
          type: number
        warrantyYears:
          type: integer
        cmcCharges:
          type: number
        cmcStartYear:
          type: integer
        status:
          type: string
          description: "active | expired | closed | renewed"
        startDate:
          type: string
        endDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - contractNumber
        - equipmentId
        - equipmentName
        - vendorId
        - vendorName
        - unitPrice
        - gstRate
        - warrantyYears
        - cmcCharges
        - cmcStartYear
        - status
        - startDate
        - endDate
        - createdAt
        - updatedAt
    CreateRateContractBody:
      type: object
      properties:
        equipmentId:
          type: integer
        vendorId:
          type: integer
        unitPrice:
          type: number
        gstRate:
          type: number
        warrantyYears:
          type: integer
        cmcCharges:
          type: number
        cmcStartYear:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
      required:
        - equipmentId
        - vendorId
        - unitPrice
        - gstRate
        - warrantyYears
        - cmcCharges
        - cmcStartYear
        - startDate
        - endDate
    UpdateRateContractBody:
      type: object
      properties:
        unitPrice:
          type: number
        gstRate:
          type: number
        status:
          type: string
        endDate:
          type: string
        cmcCharges:
          type: number

    # ─── PURCHASE ORDER ────────────────────────────────────────────────────────
...
    Tender:
      type: object
      properties:
        id:
          type: integer
        tenderNumber:
          type: string
        indentId:
          type: integer
        equipmentName:
          type: string
        status:
          type: string
          description: "invited | bids_received | l1_identified | rc_created | cancelled"
        tenderInvitedDate:
          type: ["string", "null"]
        bidsReceivedDate:
          type: ["string", "null"]
        l1BidderName:
          type: ["string", "null"]
        l1BidderAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenderNumber
        - indentId
        - equipmentName
        - status
        - tenderInvitedDate
        - bidsReceivedDate
        - l1BidderName
        - l1BidderAmount
        - notes
        - createdAt
        - updatedAt
    CreateTenderBody:
      type: object
      properties:
        indentId:
          type: integer
        tenderInvitedDate:
          type: string
        notes:
          type: string
      required:
        - indentId
        - tenderInvitedDate
    UpdateTenderBody:
      type: object
      properties:
        status:
          type: string
        bidsReceivedDate:
          type: string
        l1BidderName:
          type: string
        l1BidderAmount:
          type: number
        notes:
          type: string

    # ─── DELIVERY ─────────────────────────────────────────────────────────────
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        vendorCode:
          type: string
        name:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        isL1Bidder:
          type: boolean
        performanceScore:
          type: ["number", "null"]
        status:
          type: string
          description: "active | suspended | blacklisted"
        createdAt:
          type: string
      required:
        - id
        - vendorCode
        - name
        - contactEmail
        - contactPhone
        - address
        - gstNumber
        - isL1Bidder
        - performanceScore
        - status
        - createdAt
    CreateVendorBody:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
      required:
        - name
        - contactEmail
        - contactPhone
        - address
        - gstNumber

    # ─── INSTITUTION ──────────────────────────────────────────────────────────
...